requirements.txt
certifi==2021.5.30
chardet==3.0.4
idna==2.8
lxml
numpy
Pillow==9.3.0
pycryptodome==3.10.1
python-dateutil==2.8.2
requests==2.22.0
selenium==3.141.0
six==1.16.0
tqdm==4.48.2
urllib3==1.25.11
vobject==0.9.6.1
